+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # C6Embed Changelog
 
+* [FEATURE]: Preloading is now supported: The app will be preloaded if
+  the user mouses over the splash page, or if the data-preload attribute
+  is present on the embed tag
+
 ## v1.0.2 (June 11, 2014)
 * [FIX]: Fix for an issue where elements of the parent site would appear
   above a fullscreen experience